refactor tensorflow code
