+++ python/sqlflow_submitter/tensorflow/get_tf_version.py
@@ -1,0 +1,23 @@
+# Copyright 2020 The SQLFlow Authors. All rights reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import tensorflow as tf
+
+
+# TODO(shendiaomo): Remove after we fully upgrade to TF2.0
+def tf_is_version2():
+    is_version2 = True
+    version_parts = tf.__version__.split(".")
+    if int(version_parts[0]) == 1:
+        is_version2 = False
+    return is_version2
